refactor: week1 수정 완료 close #1
- BasicTextField -> TvingTextField 수정
- 주석 제거
- 정규식 유틸 분리
