+++ app/src/main/res/values/strings.xml
@@ -35,5 +35,6 @@
     <string name="btn_basic_topbar_back">뒤로가기</string>
     <string name="btn_password_visibility">password visibility button</string>
     <string name="btn_logout">로그아웃</string>
+    <string name="success_signin_message">회원가입 성공!</string>
 
 </resources>
